Add files via upload
MSI installer and the source code for the 3 main forms. Note that the installer includes the images folder which is referenced in the MainFrom.cs file by a relative path. Seems to work on my machine if I install it in a different folder, I hope it will find success on other systems.