Add files via upload
I figured it would be important to include the files for the designer cs files as well. I think all the names of the elements are in here. Bear in mind this code was created using Visual Studio, I didn't write this code by hand. Most of the information is there to construct the form. I think this simultaneously showcases some early strengths and weaknesses when doing frontend development. The spacing of the elements needs work, and the graphical elements leave much to be desired. However, it is all functional. That has more to do with how this ties into the backend side of this project, the main thing I wanted to demonstrate. This is only one of a handful of independent projects I've worked on, and one of the ones that was the closest to being "complete". I consider it post v1.0 as the core functionality that was designed for it is there. I will add more features and maybe a companion app. Just have to figure out how to implement csv files. It's the next best thing to HL7 which is something I'm interested in learning.