Minor updates
I re-enabled the radio buttons in the program so that users can set the day of the week manually. There was also an improvement in the control structures, it still uses switch case statements based on the day of the week, but now there's actual event handlers which set it, making it possible to retain the switch case structure I designed instead of using messy nested if statements.

Among other things, I also added reset buttons, and a way to get average totals. This is mainly a test for what will eventually become the running total feature, and a couple menu items were added to the UI as placeholder; Export and Running Total. Export will convert the information from the the timers into a csv file. Running Totals will take the information from the csv files and get an average from multiple weeks of using the timer.

Say for instance you want to keep track of the average number of hours spent studying over a 9-week period. This feature will take each line in the csv file and get the weekly totals and weekly averages for each line in the csv file. I'm still unclear of the implementation, but I will have time to work on that in the future. For now those buttons are disabled, and the logic and forms for those features have not been created. Yet. Once this version is completed it will be available on my website for download.